Add option to limit concurrent helm calls
Have added a new argument `--max-connections` that can be used to set a
limit on the number of concurrent calls to helm at one time. To enable
this I have switched `SyncCharts` from using a WaitGroup to using a pool
of workers and a queue of jobs. To ensure that this is thread safe and
an attempt is made to sync each chart at the end.

Fixes https://github.com/roboll/helmfile/issues/23