Do not try to load env values in base helmfile
When helmfile is run with `--environment NAME` and there was a base hemlfile that misses `environments`, helmfile had been trying to load env values for NAME and failing. We should skip this completely, as the env values should be loaded after merging the base helmfile into the parent helmfile.

Fixes #1008