fix: Parse --set flag for write-values
Currently, the `--set` flag doesn't do anything when passed to the
`write-values` subcommand. This uses the `strvals` package from Helm to
parse the flag values as they would be by Helm and merges the resulting
YAML into the generated values.

The use of temporary files makes it difficult to test the write-values
subcommand with a unit test, so I've added an integration test instead.