Add ability to control skipping Deps and Repos separately
Right now we have only one flag which passed to both variables, which isn't flexible as we would like to.

Main problem in our case that we are using Helmfile with envs. In each env we can have different version of remote chart. To apply that configuration in parallel we would like to split repos sync and deps build into 2 stages.

Repos sync can't be executed in parralel because it cause helm locks and `context deadline exceeded` exceptions.