Support templated secret values (#1700)
Secret files ending with .gotmpl are now also rendered as template.
For that I had to fix the assumed filename of the decrypted secret file, because it predicted it wrong as soon as the file did not end with .yaml. After that I had to make the handler method for decrypting files (generateSecretValuesFiles) use the same template processing method as for normal value files (generateTemporaryReleaseValuesFiles).
A new unit test has been added and the existing integration test has been extended.